handlers

+++ internal/domain/models/product.go
@@ -10,7 +10,7 @@ type Product struct {
 	Id             uuid.UUID `db:"id"`
 	Name           string    `db:"name"`
 	Category       string    `db:"category"`
-	Price          float32   `db:"price"`
+	Price          float64   `db:"price"`
 	AvailableStock int       `db:"available_stock"`
 	LastUpdateDate time.Time `db:"last_update_date"`
 	SupplierId     uuid.UUID `db:"supplier_id"`
